requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2021.10.8
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
intel-openmp==2022.0.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.16.4
opencv-python==4.5.5.62
pandoc==2.0.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
plumbum==1.7.2
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.14.11
pytesseract==0.3.0
python-dateutil==2.8.0
pytz==2019.2
pywin32==303
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
Send2Trash==1.5.0
sip==4.19.8
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
